# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Unlock Content Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendation/analyze:
    post:
      operationId: analyzeTranscript
      tags: [recommendation]
      summary: "Pass 1: Analyze transcript for persona, stage, objections"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeTranscriptBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptAnalysis"
  /recommendation/rank:
    post:
      operationId: rankDocuments
      tags: [recommendation]
      summary: "Pass 2: Rank eligible documents"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RankDocumentsBody"
      responses:
        "200":
          description: Ranked recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResult"
  /recommendation/confirm-send:
    post:
      operationId: confirmSend
      tags: [recommendation]
      summary: Log a send action against a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmSendBody"
      responses:
        "200":
          description: Send logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendLogEntry"
  /recommendation/email-draft:
    post:
      operationId: generateEmailDraft
      tags: [recommendation]
      summary: Generate a covering email draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDraftBody"
      responses:
        "200":
          description: Email draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailDraft"
...
components:
  schemas:
...
    SendLogEntry:
      type: object
      properties:
        send_id:
          type: string
        date:
          type: string
        documents_sent:
          type: array
          items:
            type: string
        pipeline_stage_at_send:
          type: string
        transcript_summary:
          type: string
        analysis_confidence:
          type: object
          properties:
            persona:
              type: number
            stage:
              type: number
          required:
            - persona
            - stage
        already_sent_exclusions:
          type: array
          items:
            type: string
        blocked_documents_seen:
          type: array
          items:
            type: string
        stage_suggestion:
          type: ["string", "null"]
        email_sent:
          type: boolean
        status:
          type: string
      required:
        - send_id
        - date
        - documents_sent
        - pipeline_stage_at_send
        - transcript_summary
        - analysis_confidence
        - already_sent_exclusions
        - blocked_documents_seen
        - email_sent
        - status
...
    AnalyzeTranscriptBody:
      type: object
      properties:
        transcript:
          type: string
        lead_id:
          type: ["string", "null"]
      required:
        - transcript
    TranscriptAnalysis:
      type: object
      properties:
        detected_persona:
          type: object
          properties:
            name:
              type: string
            confidence_score:
              type: number
            evidence:
              type: array
              items:
                type: string
          required:
            - name
            - confidence_score
            - evidence
        pipeline_stage:
          type: object
          properties:
            stage:
              type: string
            confidence_score:
              type: number
            rationale:
              type: string
          required:
            - stage
            - confidence_score
            - rationale
        objections:
          type: array
          items:
            $ref: "#/components/schemas/Objection"
        transcript_summary:
          type: string
        pipeline_stage_suggestion:
          type: ["string", "null"]
      required:
        - detected_persona
        - pipeline_stage
        - objections
        - transcript_summary
    Objection:
      type: object
      properties:
        objection:
          type: string
        suggested_response:
          type: string
      required:
        - objection
        - suggested_response
    RankDocumentsBody:
      type: object
      properties:
        lead_id:
          type: ["string", "null"]
        detected_persona:
          type: string
        pipeline_stage:
          type: string
        transcript_summary:
          type: string
        objections:
          type: array
          items:
            type: string
      required:
        - detected_persona
        - pipeline_stage
        - transcript_summary
    RecommendationResult:
      type: object
      properties:
        ranked_documents:
          type: array
          items:
            $ref: "#/components/schemas/RankedDocument"
        already_sent:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSummary"
        blocked_documents:
          type: array
          items:
            $ref: "#/components/schemas/BlockedDocument"
        all_sent_message:
          type: ["string", "null"]
      required:
        - ranked_documents
        - already_sent
        - blocked_documents
    RankedDocument:
      type: object
      properties:
        document_id:
          type: string
        file_code:
          type: string
        name:
          type: string
        description:
          type: string
        tier:
          type: integer
        priority:
          type: integer
        rationale:
          type: string
      required:
        - document_id
        - file_code
        - name
        - description
        - tier
        - priority
        - rationale
    DocumentSummary:
      type: object
      properties:
        document_id:
          type: string
        file_code:
          type: string
        name:
          type: string
        date_sent:
          type: string
      required:
        - document_id
        - file_code
        - name
        - date_sent
    BlockedDocument:
      type: object
      properties:
        document_id:
          type: string
        file_code:
          type: string
        name:
          type: string
        reason:
          type: string
      required:
        - document_id
        - file_code
        - name
        - reason
    ConfirmSendBody:
      type: object
      properties:
        lead_id:
          type: string
        document_ids:
          type: array
          items:
            type: string
        transcript_summary:
          type: string
        detected_persona:
          type: string
        pipeline_stage:
          type: string
        analysis_confidence:
          type: object
          properties:
            persona:
              type: number
            stage:
              type: number
          required:
            - persona
            - stage
        stage_suggestion:
          type: ["string", "null"]
        email_sent:
          type: boolean
        already_sent_exclusions:
          type: array
          items:
            type: string
        blocked_documents_seen:
          type: array
          items:
            type: string
      required:
        - lead_id
        - document_ids
        - transcript_summary
        - detected_persona
        - pipeline_stage
        - analysis_confidence
        - email_sent
    EmailDraftBody:
      type: object
      properties:
        lead_name:
          type: string
        detected_persona:
          type: string
        pipeline_stage:
          type: string
        document_names:
          type: array
          items:
            type: string
        transcript_summary:
          type: string
      required:
        - lead_name
        - detected_persona
        - pipeline_stage
        - document_names
        - transcript_summary
    EmailDraft:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        template_used:
          type: string
      required:
        - subject
        - body
        - template_used
...