# lib/api-zod/src/generated/types/analyzeTranscriptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unlock Content Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeTranscriptBody {
  transcript: string;
  /** @nullable */
  lead_id?: string | null;
}