# lib/api-zod/src/generated/types/objection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unlock Content Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Objection {
  objection: string;
  suggested_response: string;
}